docs: Tell users that pip is the best way to install ansible
Ansible is hell to install through the operating system. Its
dependenices (like sshpass) are a good idea to install through
a package manager, but installing Ansible itself that way
risks instability. It'd ultimately always be preferable to
provide a requirements.txt specifying which Ansible version
should be installed in order to keep the playbooks working.